Disable require-default-props lint rule
The require-default-props lint rule was being overly restrictive for
this use case. There is no good reason to include a value in
`defaultProps` if it is truly optional, as that just means that React
now has to do the work to merge this `undefined` value into props for no
reason.

I'd argue that you should just avoid `defaultProps` all together and use
destructuring with default values, but that's maybe more personal
preference.