Card number should be visible
The example code for React Hooks Card Detailed  sample currently sets the foreground color for the card number to white. Setting the foreground color to white on a white background field makes it impossible to see the number when entering it in the form. 

The correct foreground color value to use in the example is a dark color like black.